backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
psycopg2-binary==2.9.6
langchain==0.0.352
openai==1.3.7
python-dotenv==1.0.0